refactor polygon to models module (#135)
