+++ csep/core/regions.py
@@ -5,14 +5,14 @@ from itertools import compress
 from xml.etree import ElementTree as ET
 
 # Third-party imports
-import matplotlib.path
 import numpy
 import numpy as np
-import pyproj
 
 # PyCSEP imports
 from csep.utils.calc import bin1d_vec, cleaner_range, first_nonnan, last_nonnan
 from csep.utils.scaling_relationships import WellsAndCoppersmith
+from csep.models import Polygon
+
 
 def california_relm_collection_region(dh_scale=1, magnitudes=None, name="relm-california-collection"):
     """ Return collection region for California RELM testing region
@@ -463,74 +463,6 @@ def _bin_catalog_probability(lons, lats, n_poly, mask, idx_map, binx, biny):
     event_counts[hash_idx] = 1
     return event_counts
 
-class Polygon:
-    """
-    Represents polygons defined through a collection of vertices.
-
-    This polygon is assumed to be 2d, but could contain an arbitrary number of vertices. The path is treated as not being
-    closed.
-    """
-    def __init__(self, points):
-        # instance members
-        self.points = points
-        self.origin = self.points[0]
-
-        # https://matplotlib.org/3.1.1/api/path_api.html
-        self.path = matplotlib.path.Path(self.points)
-
-    def __str__(self):
-        return str(self.origin)
-
-    def contains(self, points):
-        """ Returns a bool array which is True if the path contains the corresponding point.
-
-        Args:
-            points: 2d numpy array
-
-        """
-        nd_points = np.array(points)
-        if nd_points.ndim == 1:
-            nd_points = nd_points.reshape(1,-1)
-        return self.path.contains_points(nd_points)
-
-    def centroid(self):
-        """ return the centroid of the polygon."""
-        c0, c1 = 0, 0
-        k = len(self.points)
-        for p in self.points:
-            c0 = c0 + p[0]
-            c1 = c1 + p[1]
-        return c0 / k, c1 / k
-
-    def get_xcoords(self):
-        return np.array(self.points)[:,0]
-
-    def get_ycoords(self):
-        return np.array(self.points)[:,1]
-
-    @classmethod
-    def from_great_circle_radius(cls, centroid, radius, num_points=10):
-        """
-        Generates a polygon object from a given radius and centroid location.
-
-        Args:
-            centroid: (lon, lat)
-            radius: should be in (meters)
-            num_points: more points is higher resolution polygon
-
-        Returns:
-            polygon
-        """
-        geod = pyproj.Geod(ellps='WGS84')
-        azim = np.linspace(0, 360, num_points)
-        # create vectors with same length as azim for computations
-        center_lons = np.ones(num_points) * centroid[0]
-        center_lats = np.ones(num_points) * centroid[1]
-        radius = np.ones(num_points) * radius
-        # get new lons and lats
-        endlon, endlat, backaz = geod.fwd(center_lons, center_lats, azim, radius)
-        # class method
-        return cls(np.column_stack([endlon, endlat]))
 
 class CartesianGrid2D:
     """Represents a 2D cartesian gridded region.
